+++ CHANGES.md
@@ -1,3 +1,9 @@
+1.5.2
+=====
+
+* Add `logPureAction` to `PureLogging`.
+* Add `withSublogger` to `HasLoggerName`.
+
 1.5.1
 =====